Took calculate BF and RU out of github-wrapper.ts
