+++ src/github-wrapper.ts
@@ -8,89 +8,12 @@ import * as API from './api-calls/github-adapter';
 import { calculateCorrectness } from './find-correctness';
 import { calculateResponsiveMaintener } from './find-responsive-maintainer';
 import { calculateNetScore } from './netscore';
+import { calculateBusFactor } from './bus-factor';
+import { calculateRampUp } from './ramp-up'
 
 
 if (!Util.Constants.GITHUB_TOKEN) {
   Util.Logger.logErrorAndExit('Error: GITHUB_TOKEN is not set in the environment.');
-}
-
-/**
- * Fetch repository details from GitHub to calculate the Ramp-Up score.
- * @param owner Repository owner (username or organization)
- * @param repo Repository name
- */
-async function fetchRepoDetails(owner: string, repo: string) {
-  try {
-    const response = await axios.get(`${Util.Constants.GITHUB_API_BASE_URL}/repos/${owner}/${repo}`, {
-      headers: {
-        Authorization: `token ${Util.Constants.GITHUB_TOKEN}`,
-      },
-    });
-    return response.data;
-  } catch (error) {
-    console.error(`Error fetching repository details: ${error}`);
-    throw error;
-  }
-}
-
-/**
- * Calculate Ramp-Up score for the repository based on several factors.
- * The score is based on the presence of a README file, open issues, and the presence of a docs folder.
- * @param owner Repository owner (username or organization)
- * @param repo Repository name
- */
-async function calculateRampUp(owner: string, repo: string): Promise<number> {
-  try {
-    const repoDetails = await fetchRepoDetails(owner, repo);
-
-    let rampUpScore = 0;
-
-    // Check if README.md exists (5 points if present)
-    const hasReadme = repoDetails.has_wiki || repoDetails.description;
-    if (hasReadme) {
-      rampUpScore += 5;  // README is present
-    }
-
-    // Consider the number of open issues (fewer issues = better ramp-up score)
-    const openIssues = repoDetails.open_issues_count;
-    if (openIssues < 10) {
-      rampUpScore += 5;  // Very few open issues
-    } else if (openIssues < 50) {
-      rampUpScore += 3;  // Moderate number of open issues
-    } else {
-      rampUpScore += 1;  // High number of open issues
-    }
-
-    return rampUpScore;
-
-  } catch (error) {
-    console.error(`Error calculating Ramp-Up score for ${owner}/${repo}: ${error}`);
-    return 0;
-  }
-}
-
-/**
- * Calculate Bus Factor
- * @param contributors List of contributors with commit count
- * @param threshold Contribution threshold for bus factor (default 50%)
- */
-function calculateBusFactor(contributors: any[], threshold: number = 50): number {
-  const totalCommits = contributors.reduce((acc, contributor) => acc + contributor.contributions, 0);
-  const sortedContributors = contributors.sort((a, b) => b.contributions - a.contributions);
-
-  let commitSum = 0;
-  let busFactor = 0;
-
-  for (const contributor of sortedContributors) {
-    commitSum += contributor.contributions;
-    busFactor += 1;
-    const contributionPercentage = (commitSum / totalCommits) * 100;
-    if (contributionPercentage >= threshold) {
-      break;
-    }
-  }
-
-  return busFactor;
 }
 
 /**
@@ -133,15 +56,15 @@ async function calculateMetricsForRepo(githubUrl: string): Promise<string> {
 
     const correctness = await calculateCorrectness(owner, repo);
 
-      const maintainResponsiveness = await calculateResponsiveMaintener(owner, repo);
+    const maintainResponsiveness = await calculateResponsiveMaintener(owner, repo);
 
-      // Calculate NetScore
-      const netScore = calculateNetScore(
+    // Calculate NetScore
+    const netScore = calculateNetScore(
           busFactor.output,
           rampUpScore.output,
           correctness,
           maintainResponsiveness
-      );
+     );
 
     //var ndjsonOutputString = busFactor.output;
